requirements.txt
fastapi
sqlalchemy
uvicorn
python-dotenv
email-validator
werkzeug
python-multipart
python-jose[cryptography]
